+++ src/selectors/articles/select_articles_for_explore_detail.ts
@@ -5,11 +5,11 @@ import { selectLocale } from '../locale';
 import { RouterProps } from '../../application/routing';
 import { toSelectorArticleListItem } from './to_selector_article_list_item';
 import { ArticleListItem } from './types';
-import { pullExploreSection } from '../explore/pull_explore_section';
+import { pullExploreSectionById } from '../explore/pull_explore_section_by_id';
 
 export const selectArticlesForExploreDetail = (store: Store, routerProps: RouterProps): ReadonlyArray<ArticleListItem> => {
     const exploreId = routerProps.match.params.learnId;
-    const exploreSection = pullExploreSection(store, exploreId);
+    const exploreSection = pullExploreSectionById(store, exploreId);
     const articles = store.articlesInStore.articles;
     const matchingArticles = taskDetails.findItemByLearnTaxonomyTerm(exploreSection.taxonomyTerms, articles);
