+++ src/selectors/explore/pull_explore_section_by_id.ts
@@ -1,6 +1,6 @@
 import { Store } from '../../stores';
 import * as model from '../../stores/explore';
 
-export const pullExploreSection = (store: Store, learnId: model.Id): model.ExploreSection => (
+export const pullExploreSectionById = (store: Store, learnId: model.Id): model.ExploreSection => (
     store.exploreSectionsInStore.sections[learnId]
 );
